Use float32 for the cache

+++ nnpz/python/nnpz/reference_sample/PdzDataProvider.py
@@ -133,7 +133,7 @@ class PdzDataProvider(object):
         if pos > self.__entries:
             raise InvalidAxisException('Position {} for {} entries'.format(pos, self.__entries))
         if self.__cache is None:
-            self.__cache = np.zeros((self.__cache_size, len(self.__redshift_bins)))
+            self.__cache = np.zeros((self.__cache_size, len(self.__redshift_bins)), dtype=np.float32)
         cache_line, cache_offset = divmod(pos, self.__cache_size)
         if cache_line != self.__cache_line:
             cache_start = cache_line * self.__cache_size
